Move write_stubs to Namespace
