+++ tierkreis/tests/idl/test_idl.py
@@ -5,7 +5,6 @@ from tierkreis.idl.models import GenericType
 from tierkreis.namespace import Namespace
 from tierkreis.idl.type_symbols import type_symbol
 import tests.idl.namespace1
-from tierkreis.worker.worker import Worker
 
 type_symbols = [
     ("uint8", GenericType(int, [])),
@@ -46,9 +45,7 @@ def test_namespace(path: Path, expected: Namespace):
     # Write stubs to file.
     # This file will be subject to linting.
     # Also a change in this file can indicate an unexpectedly breaking change.
-    worker = Worker("dummy_worker")
-    worker.namespace = namespace
-    worker.write_stubs(Path(__file__).parent / "stubs_output.py")
+    namespace.write_stubs(Path(__file__).parent / "stubs_output.py")
 
 
 @pytest.mark.parametrize("type_symb", type_symbols_for_failure)
